+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Changelog
 
 ## TBD
 
+* Adds the `Telemetry` configuration property.
 * Updates the bugsnag-android dependency from v5.22.1 to [v5.23.0](https://github.com/bugsnag/bugsnag-android/blob/master/CHANGELOG.md#5230-2022-06-20)
 * Updates the bugsnag-cocoa dependency from v6.16.8 to [v6.19.0](https://github.com/bugsnag/bugsnag-cocoa/blob/master/CHANGELOG.md#6190-2022-06-29)